Add files via upload
Script to get prices on close of Friday and beginning of monday for the past year. Used to see if it is a better investment strategy to sell AMD stock at the end of the week and rebuy it on open the next Monday. Used to see if overall change over the weekend is negative or positive over a long period of time